p29.31 refactor: thread ids.AccountID everywhere
Convert the account id from bare int64 to the distinct ids.AccountID type
across the whole stack (the last + biggest of the p29 typed-id capstone):
sqlc models/queries, store (accounts, options, transactions, merge, balances,
history, imports, expense reports, budget plans, reconciliations, last_used),
reports (compute/drill/income-statement/balance-sheet/program/trial-balance/
form-990/fund-activity/budget-plan/capital-campaign), web handlers, synth,
devseed, ledgerimport, and the fixture. Bare int64 remains only at the SQL
scan boundary (sqlc query funcs) and the HTTP parse boundary (ids.AccountID
(parseID(s))); nullable parent FKs stay sql.NullInt64 in the model and are
typed at the store boundary. Removed the now-dead nullInt64Ptr helper.

Purely a type change: report goldens are byte-identical (no testdata
regenerated), make lint/test/check all green.

+++ cmd/ledgerimport/build_test.go
@@ -1737,10 +1737,10 @@ func TestCorrectionPostsBalancedAdjustment(t *testing.T) {
 	for _, l := range legs {
 		byAcct[l.acct] = l
 	}
-	if l, ok := byAcct[res.AccountIDs["Checking"]]; !ok || l.amount != 400000 {
+	if l, ok := byAcct[int64(res.AccountIDs["Checking"])]; !ok || l.amount != 400000 {
 		t.Errorf("Checking leg = %+v, want amount 400000", l)
 	}
-	if l, ok := byAcct[res.AccountIDs["Grant Revenue"]]; !ok || l.amount != -400000 {
+	if l, ok := byAcct[int64(res.AccountIDs["Grant Revenue"])]; !ok || l.amount != -400000 {
 		t.Errorf("Grant Revenue leg = %+v, want amount -400000", l)
 	} else if !l.prog.Valid || l.prog.Int64 != int64(res.ProgramIDs["Education"]) {
 		t.Errorf("Grant Revenue leg program = %v, want Education %d", l.prog, res.ProgramIDs["Education"])
